requirements.txt
matplotlib
numpy
pandas
prettytable
torch
PyYAML
tensorboard
